+++ integration-tests/library/src/main/java/com/squareup/anvil/test/Scopes.kt
@@ -1,0 +1,5 @@
+package com.squareup.anvil.test
+
+abstract class AppScope private constructor()
+
+abstract class SubScope private constructor()