+++ sample/app/src/main/java/com/squareup/anvil/sample/LoggedInComponent.kt
@@ -1,0 +1,9 @@
+package com.squareup.anvil.sample
+
+import com.squareup.anvil.annotations.MergeSubcomponent
+import com.squareup.scopes.LoggedInScope
+import com.squareup.scopes.SingleIn
+
+@SingleIn(LoggedInScope::class)
+@MergeSubcomponent(LoggedInScope::class)
+interface LoggedInComponent