+++ sample/app/src/main/java/com/squareup/anvil/sample/RandomModule.kt
@@ -1,0 +1,18 @@
+package com.squareup.anvil.sample
+
+import com.squareup.anvil.annotations.ContributesTo
+import com.squareup.scopes.AppScope
+import dagger.Module
+import dagger.Provides
+import javax.inject.Named
+
+@Module
+@ContributesTo(AppScope::class)
+object RandomModule {
+  @Provides @Named("random") fun provideString(): String = "Hello!"
+}
+
+@ContributesTo(AppScope::class)
+interface RandomComponent {
+  @Named("random") fun string(): String
+}