+++ sample/app/src/main/java/com/squareup/hephaestus/sample/LoggedInComponent.kt
@@ -1,9 +1,0 @@
-package com.squareup.hephaestus.sample
-
-import com.squareup.hephaestus.annotations.MergeSubcomponent
-import com.squareup.scopes.LoggedInScope
-import com.squareup.scopes.SingleIn
-
-@SingleIn(LoggedInScope::class)
-@MergeSubcomponent(LoggedInScope::class)
-interface LoggedInComponent