+++ sample/app/src/test/java/com/squareup/anvil/sample/LoggedInComponentTest.kt
@@ -1,0 +1,14 @@
+package com.squareup.anvil.sample
+
+import com.google.common.truth.Truth.assertThat
+import org.junit.Test
+
+class LoggedInComponentTest {
+
+  @Test fun `LoggedInComponent can be created`() {
+    val loggedInComponent = DaggerAppComponent.create()
+        .loggedInComponent()
+
+    assertThat(loggedInComponent).isNotNull()
+  }
+}