+++ sample/library/src/main/java/com/squareup/anvil/sample/mother/MotherProvider.kt
@@ -1,0 +1,7 @@
+package com.squareup.anvil.sample.mother
+
+import com.squareup.anvil.sample.God
+
+interface MotherProvider {
+  fun mother(god: God): String
+}