+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Version
 
+* Renamed the project from Hephaestus to Anvil #12. **IMPORTANT:** Anvil is not compatible with Hephaestus and you must upgrade the plugin in all of your libraries.
+
 ## 1.0.4-1.4-M3 (2020-07-24)
 
 * Ensure that types cannot be included and excluded at the same time, which leads to unexpected results.